% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airfoil.R
\docType{data}
\name{airfoil}
\alias{airfoil}
\title{Airfoil Self-Noise Data by NASA}
\format{
This problem has the following inputs:
\enumerate{
\item Frequency, in Hertzs.
\item Angle of attack, in degrees.
\item Chord length, in meters.
\item Free-stream velocity, in meters per second.
\item Suction side displacement thickness, in meters.
}

The only output is:
\enumerate{
\item Scaled sound pressure level, in decibels.
}
}
\usage{
airfoil
}
\description{
The NASA data set comprises different size NACA 0012 airfoils at various wind tunnel speeds and angles of attack. The span of the airfoil and the observer position were the same in all of the experiments.
}
\author{
Brooks,Thomas, Pope,D., and Marcolini,Michael. (2014). Airfoil Self-Noise. UCI Machine Learning Repository. https://doi.org/10.24432/C5VW2C.
}
\keyword{datasets}
