% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allbus.R
\docType{data}
\name{allbus}
\alias{allbus}
\title{Allbus 2021}
\format{
544 Variablen mit 5342 Beobachtungen

Eine Liste aller Variablen kann in der \code{dictionary}-Tabelle nachgesehen werden

Metrische Variablen:
\itemize{
\item \strong{age}:
Alter
\item \strong{di05}:
Haushaltsnettoeinkommen
\item \strong{lm02}:
Fernsehgesamtdauer pro Tag in Minuten
}
}
\source{
https://www.gesis.org/allbus/download
}
\usage{
allbus
}
\description{
Allgemeine Bevölkerungsumfrage der Sozialwissenschaften
}
\keyword{datasets}
