% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cranpacks.R
\docType{data}
\name{cranpacks}
\alias{cranpacks}
\title{cranpacks}
\format{
4 Variablen mit 19995 Beobachtungen
\itemize{
\item \strong{id}: fortlaufende Nummer (zählt nur Pakete, die zum Stand veröffentlicht sind)
\item \strong{date}: Datum der Aufnahme in CRAN
\item \strong{package}: Paketname
\item \strong{title}: Titel des Pakets
}
}
\source{
abgerufen von https://cran.r-project.org/web/packages/available_packages_by_date.html
}
\usage{
cranpacks
}
\description{
Liste der auf CRAN veröffentlichten Pakete, Stand 25.10.2023
}
\keyword{datasets}
