% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dict.R
\docType{data}
\name{dict}
\alias{dict}
\title{Codebook zum \code{allbus}-Datensatz}
\format{
7 Spalten und 6930 Zeilen

Eine Liste aller Variablen und den Bezeichnungen ihrer Ausprägungen

Das Label \verb{[-32] NICHT GENERIERBAR} bezeichnet metrische Ausprägungen

Die Liste ist im Langformat, d.h. zu einer Variable können mehrere Spalten
gehören wenn sie mehrere Ausprägungen besitzt.
\itemize{
\item \strong{pos}:
Position
\item \strong{variable}:
in R nutzbarerer Name
\item \strong{label}:
Titel
\item \strong{col_type}:
Datentyp
\item \strong{missing}:
Anzahl fehlender Werte
\item \strong{value-labels}:
Codierungen
}
}
\usage{
dict
}
\description{
(dictionary)
}
\details{
Daten: Allbus 2021

Allgemeine Bevölkerungsumfrage der Sozialwissenschaften
}
\keyword{datasets}
