% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net.r
\docType{data}
\name{net}
\alias{net}
\title{Neun-Euro-Ticket Umfrage 2022}
\format{
2 Variablen mit 2446 Beobachtungen

Metrische Variablen:
\itemize{
\item \strong{alter}:
Alter
\item \strong{nutzung}:
Nutzung des Tickets an X Tagen pro Monat
}
}
\source{
https://www.uni-kassel.de/fb14bau/institute/institut-fuer-verkehrswesen-ifv/radverkehr-und-nahmobilitaet/forschung/neun-euro-ticket
}
\usage{
net
}
\description{
Umfrage Anlässlich des in Deutschland eingeführten 9-Euro-Tickets
}
\keyword{datasets}
