% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publictr.R
\docType{data}
\name{publictr}
\alias{publictr}
\title{public transport}
\format{
6 Variablen und 456 Beobachtungen
\itemize{
\item Jahr
\item Quartal
\item Typ (siehe Anmerkungen)
\item Personen (beförderte Personen)
\item Personenkilometer
\item ausgewertete Unternehmen
}

\strong{Anmerkungen zur Variable Typ}

\describe{
\item{Liniennahverkehr insgesamt:}{Unternehmensfahrten}
\item{Liniennahverkehr mit Eisenbahnen, Straßenbahnen und Omnibussen:}{Verkehrsmittelfahrten}
\item{Linienfernverkehr mit Omnibussen:}{Nach der Liberalisierung des Linienfernverkehrs mit Omnibussen zum 1. Januar 2013 werden die neugegründeten Unternehmen aus rechtlichen Gründen erst ab dem 1. Quartal 2016 in die Quartalserhebung einbezogen. Um Fehlinterpretationen zu vermeiden, wurden daher bis zum 4. Quartal 2015 keine Ergebnisse für den Omnibus-Linienfernverkehr nachgewiesen}
}
}
\usage{
publictr
}
\description{
Daten des statistischen Bundesamts über die Nutzung von Bussen und Bahnen in Deutschland von 2004 bis 2022
}
\references{
\href{https://www-genesis.destatis.de/genesis//online?operation=table&code=46181-0005&bypass=true&levelindex=0&levelid=1659561334845#abreadcrumb}{Tabelle 46181-00005}
}
\author{
© Statistisches Bundesamt (Destatis), 2022 | Stand: 04.08.2022
}
\keyword{datasets}
