% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/einkaufen.R
\docType{data}
\name{einkaufen}
\alias{einkaufen}
\title{einkaufen}
\format{
12 Variablen mit 166 Beobachtungen
\itemize{
\item \strong{befragung}:
Initialien der Studierenden, die die Befragung durchgeführt haben
\item \strong{id}:
Interviewnummer für jede einzelne Befragung
\item \strong{ort}
\item \strong{bioladen}:
TRUE / FALSE, je nach dem ob der Laden ein Bioladen war
\item \strong{datum}: YYYY-MM-DD
\item \strong{uhrzeit}
\item \strong{essen}: Lieblingsessen
\item \strong{kochen}: Anzahl der Tage pro Woche, an denen selber gekocht wird
\item \strong{verkehrsmittel}: Wie sind Sie hier her gekommen?
\item \strong{weg}: Wie weit war ihr Weg hier her? (in km)
\item \strong{frequenz}: Wie oft gehen Sie in einen Bioladen?
\item \strong{alter}
}
}
\usage{
einkaufen
}
\description{
Von Studierenden durchgeführte Befragungen von Kund_innen in
Bioläden in und um Kassel, es sind aber auch zwei Nicht-Bioläden dabei.
}
\author{
Studierende des WiSe 22/23 in Kassel
}
\keyword{datasets}
