% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/einkaufen.R
\docType{data}
\name{einkaufen}
\alias{einkaufen}
\title{einkaufen}
\format{
9 Variablen mit 166 Beobachtungen
\itemize{
\item \strong{Befragung}:
Initialien der Studierenden, die die Befragung durchgeführt haben
\item \strong{ID}:
Interviewnummer für jede einzelne Befragung
\item \strong{Ort}
\item \strong{Bioladen}:
TRUE / FALSE, je nach dem ob der Laden ein Bioladen war
\item \strong{Datum}: YYYY-MM-DD
\item \strong{Uhrzeit}
\item \strong{Essen}: Lieblingsessen
\item \strong{Kochen}: Anzahl der Tage pro Woche, an denen selber gekocht wird
\item \strong{Verkehrsmittel}: Wie sind Sie hier her gekommen?
\item \strong{Weg}: Wie weit war ihr Weg hier her? (in km)
\item \strong{Frequenz}: Wie oft kochen Sie selbst?
\item \strong{Alter}
}
}
\usage{
einkaufen
}
\description{
Von Studierenden durchgeführte Befragungen von Kund_innen
(hauptsächlich vor Bioläden in und um Kassel)
}
\author{
Studierende des WiSe 22/23 in Kassel
}
\keyword{datasets}
